+++ packages/core/src/index.ts
@@ -50,7 +50,7 @@ export {
 } from "./evaluate.js";
 export { resolveEase, lerpValue, isColor, EASE_NAMES } from "./interpolate.js";
 export { sampleBehavior } from "./behaviors.js";
-export { collectImageSrcs } from "./assets.js";
+export { collectImageSrcs, collectVideoSrcs } from "./assets.js";
 export {
   sketchToTimeline,
   type MotionSketch,